Added resource file where are stored local path of configuration server and binary node
From visual studio add the follow resources in internal directory in this way:

From Solution Explorer select Quickstarts.MyOPCServer.Config.xml right click and choose "Properties". Set Copy to Output Directory with "if there are changes". In this way the configuration file is copied in the internal directory of project.

The same instrution for Published/Quickstarts.MyOPCServer.PredefinedNodes.uanodes